requirements.txt
numpy==1.22.3
tensorflow-gpu==2.4.1
matplotlib==3.5.2
pandas==1.4.2
scikit-learn==1.0.2
setuptools==47.1.0
pytest==7.1.1
seaborn==0.11.2
opencv-python==4.5.5.64

pydantic
evidently
black
build